+++ test/engine_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:bloub_avatar/src/engine/mascot_engine.dart';
-import 'package:bloub_avatar/src/core/shape.dart';
 
 void main() {
   test('sample does not crash', () {